# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UserRoleDTO:
    properties:
      role:
        type: string
      username:
        type: string
    required:
    - role
    - username
    type: object
...
  handler.ErrorResponse:
    properties:
      message:
        example: validation failed
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 'REST API for marketplace: products, categories, cart, orders, auth
    and roles.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Marketplace API
  version: "1.0"
paths:
...
  /roles:
    post:
      consumes:
      - application/json
      description: Assign a role to a user. Requires admin role.
      parameters:
      - description: Username and role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Role granted confirmation
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Parse or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden (not admin)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Grant role
      tags:
      - Roles
  /roles/{username}:
    get:
      description: Get all roles of a user by username. Requires admin role.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user roles
      tags:
      - Roles
  /roles/user/{username}/role/{rolename}:
    delete:
      description: Remove a role from a user. Requires admin role.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Role name
        in: path
        name: rolename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role revoked confirmation
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User or role not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke role
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"