# backend/internal/transport/handler/user_role_handler.go
package handler

import (
	"context"
	"encoding/json"
	"fmt"
	"myapp/internal/domain"
	"myapp/internal/service"
	"myapp/internal/transport/dto"
	"myapp/internal/transport/middleware"
	"myapp/internal/transport/validator"
	"myapp/internal/utils"
	"net/http"
)

type UserRoleService interface {
	Create(ctx context.Context, input service.UserRoleInput) error
	Delete(ctx context.Context, input service.UserRoleInput, currentUserID int) error
	List(ctx context.Context, username string) ([]string, error)
}

type UserRoleHandler struct {
	userRoleService UserRoleService
}

func NewUserRoleHandler(userRoleService UserRoleService) UserRoleHandler {
	return UserRoleHandler{
		userRoleService: userRoleService,
	}
}

// Create godoc
//
//	@Summary		Grant role
//	@Description	Assign a role to a user. Requires admin role.
//	@Tags			Roles
//	@Accept			json
//	@Produce		json
//	@Security		BearerAuth
//	@Param			body	body		dto.UserRoleDTO		true	"Username and role"
//	@Success		200		{object}	MessageResponse		"Role granted confirmation"
//	@Failure		400		{object}	ErrorResponse		"Parse or validation error"
//	@Failure		401		{object}	ErrorResponse		"Unauthorized"
//	@Failure		403		{object}	ErrorResponse		"Forbidden (not admin)"
//	@Failure		404		{object}	ErrorResponse		"User not found"
//	@Failure		500		{object}	ErrorResponse		"Internal server error"
//	@Router			/roles [post]
func (h UserRoleHandler) Create(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var dto dto.UserRoleDTO
	if err := json.NewDecoder(r.Body).Decode(&dto); err != nil {
		utils.WriteError(ctx, w, fmt.Errorf("user role dto: %w", domain.ErrParse))
		return
	}

	if err := validator.Struct(dto); err != nil {
		utils.WriteError(ctx, w, fmt.Errorf("user role dto: %w", domain.ErrValidation))
		return
	}

	if err := h.userRoleService.Create(ctx, service.UserRoleInput{Username: dto.Username, Role: dto.Role}); err != nil {
		utils.WriteError(ctx, w, err)
		return
	}

	utils.WriteJSON(w, map[string]string{"message": "role granted"})
}

// Delete godoc
//
//	@Summary		Revoke role
//	@Description	Remove a role from a user. Requires admin role.
//	@Tags			Roles
//	@Produce		json
//	@Security		BearerAuth
//	@Param			username	path		string	true	"Username"
//	@Param			rolename	path		string	true	"Role name"
//	@Success		200			{object}	MessageResponse	"Role revoked confirmation"
//	@Failure		401			{object}	ErrorResponse		"Unauthorized"
//	@Failure		403			{object}	ErrorResponse		"Forbidden"
//	@Failure		404			{object}	ErrorResponse		"User or role not found"
//	@Failure		500			{object}	ErrorResponse		"Internal server error"
//	@Router			/roles/user/{username}/role/{rolename} [delete]
func (h UserRoleHandler) Delete(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	user, err := middleware.UserFromContext(ctx)
	if err != nil {
		utils.WriteError(ctx, w, err)
		return
	}

	if err := h.userRoleService.Delete(ctx, service.UserRoleInput{Username: r.PathValue("username"), Role: r.PathValue("rolename")}, user.UserID); err != nil {
		utils.WriteError(ctx, w, err)
		return
	}

	utils.WriteJSON(w, map[string]string{"message": "role deleted"})
}

// List godoc
//
//	@Summary		List user roles
//	@Description	Get all roles of a user by username. Requires admin role.
//	@Tags			Roles
//	@Produce		json
//	@Security		BearerAuth
//	@Param			username	path		string	true	"Username"
//	@Success		200			{array}		string			"List of roles"
//	@Failure		401			{object}	ErrorResponse	"Unauthorized"
//	@Failure		403			{object}	ErrorResponse	"Forbidden"
//	@Failure		404			{object}	ErrorResponse	"User not found"
//	@Failure		500			{object}	ErrorResponse	"Internal server error"
//	@Router			/roles/{username} [get]
func (h UserRoleHandler) List(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	roles, err := h.userRoleService.List(ctx, r.PathValue("username"))
	if err != nil {
		utils.WriteError(ctx, w, err)
		return
	}

	utils.WriteJSON(w, roles)
}